+++ src/Symfony/Component/Serializer/ArgumentResolver/UserInputResolver.php
@@ -14,26 +14,18 @@ namespace Symfony\Component\Serializer\ArgumentResolver;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ArgumentValueResolverInterface;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
-use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
-use Symfony\Component\Serializer\Annotation\Input;
-use Symfony\Component\Serializer\Exception\PartialDenormalizationException;
+use Symfony\Component\Serializer\Annotation\RequestBody;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\SerializerInterface;
-use Symfony\Component\Validator\ConstraintViolation;
-use Symfony\Component\Validator\ConstraintViolationList;
-use Symfony\Component\Validator\Exception\InputValidationFailedException;
-use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
- * Deserialize & validate user input.
- *
- * Works in duo with Symfony\Bundle\FrameworkBundle\EventListener\InputValidationFailedExceptionListener.
+ * Deserialize request body if Symfony\Component\Serializer\Annotation\RequestBody attribute is present on an argument.
  *
  * @author Gary PEGEOT <garypegeot@gmail.com>
  */
 class UserInputResolver implements ArgumentValueResolverInterface
 {
-    public function __construct(private SerializerInterface $serializer, private ?ValidatorInterface $validator = null)
+    public function __construct(private SerializerInterface $serializer)
     {
     }
 
@@ -51,47 +43,16 @@ class UserInputResolver implements ArgumentValueResolverInterface
     public function resolve(Request $request, ArgumentMetadata $argument): iterable
     {
         $attribute = $this->getAttribute($argument);
-        $context = array_merge($attribute->serializationContext, [
+        $context = array_merge($attribute->context, [
             DenormalizerInterface::COLLECT_DENORMALIZATION_ERRORS => true,
         ]);
         $format = $attribute->format ?? $request->getContentType() ?? 'json';
 
-        try {
-            $input = $this->serializer->deserialize($request->getContent(), $argument->getType(), $format, $context);
-        } catch (PartialDenormalizationException $e) {
-            if (null === $this->validator) {
-                throw new UnprocessableEntityHttpException(message: $e->getMessage(), previous: $e);
-            }
-
-            $errors = new ConstraintViolationList();
-
-            foreach ($e->getErrors() as $exception) {
-                $message = sprintf('The type must be one of "%s" ("%s" given).', implode(', ', $exception->getExpectedTypes()), $exception->getCurrentType());
-                $parameters = [];
-
-                if ($exception->canUseMessageForUser()) {
-                    $parameters['hint'] = $exception->getMessage();
-                }
-
-                $errors->add(new ConstraintViolation($message, '', $parameters, null, $exception->getPath(), null));
-            }
-
-            throw new InputValidationFailedException(null, $errors);
-        }
-
-        if ($this->validator) {
-            $errors = $this->validator->validate(value: $input, groups: $attribute->validationGroups);
-
-            if ($errors->count() > 0) {
-                throw new InputValidationFailedException($input, $errors);
-            }
-        }
-
-        yield $input;
+        yield $this->serializer->deserialize($request->getContent(), $argument->getType(), $format, $context);
     }
 
-    private function getAttribute(ArgumentMetadata $argument): ?Input
+    private function getAttribute(ArgumentMetadata $argument): ?RequestBody
     {
-        return $argument->getAttributes(Input::class, ArgumentMetadata::IS_INSTANCEOF)[0] ?? null;
+        return $argument->getAttributes(RequestBody::class, ArgumentMetadata::IS_INSTANCEOF)[0] ?? null;
     }
 }